+++ src/compas/geometry/frame.py
@@ -24,6 +24,10 @@ from .vector import Vector
 from .point import Point
 from .quaternion import Quaternion
 
+from scipy.spatial.transform import Rotation as R
+from scipy.spatial.transform import Slerp
+from ..itertools import linspace
+
 
 class Frame(Geometry):
     """A frame is defined by a base point and two orthonormal base vectors.
@@ -591,6 +595,73 @@ class Frame(Geometry):
     # ==========================================================================
     # Methods
     # ==========================================================================
+
+    def interpolate_frame(self, other, t):
+        """Interpolates between two frames at a given parameter t in the range [0, 1]
+
+        Parameters
+        ----------
+        other : :class:`compas.geometry.Frame`
+        t: float
+            a parameter in the range [0-1]
+
+        Returns
+        -------
+        :class:`compas.geometry.Frame`
+            the interpolated frame
+
+        Examples
+        --------
+        >>> frame1 = Frame(Point(0, 0, 0), Vector(1, 0, 0), Vector(0, 1, 0))
+        >>> frame2 = Frame(Point(1, 1, 1), Vector(0, 0, 1), Vector(0, 1, 0))
+        >>> start_frame = frame1.interpolate_frame(frame2, 0)
+        >>>> allclose(start_frame.point, frame1.point) and allclose(start_frame.quaternion, frame1.quaternion)
+        True
+        """
+        quat1 = Quaternion.from_frame(self)
+        quat2 = Quaternion.from_frame(other)
+        key_rotations = R.from_quat([quat1, quat2])
+
+        key_times = [0.0, 1.0]
+
+        slerp = Slerp(key_times, key_rotations)
+
+        interpolated_rot = slerp([t])[0]
+
+        # Interpolate origin
+        origin_interpolated = (self.point * (1.0 - t)) + other.point * t
+
+        # Convert interpolated quaternion back to a compas Quaternion
+        rot = Quaternion(*interpolated_rot.as_quat().tolist())
+
+        # Create a new frame with the interpolated position and orientation
+        interpolated_frame = Frame.from_quaternion(rot, origin_interpolated)
+
+        return interpolated_frame
+
+    def interpolate_frames(self, other, steps):
+        """Generates a specified number of interpolated frames between two given frames
+
+        Parameters
+        ----------
+        other: Frame
+        steps: int
+            the number of interpolated frames to return
+
+        Returns
+        -------
+        list of :class:`compas.geometry.Frame`
+
+        Examples
+        --------
+        >>> frame1 = Frame(Point(0, 0, 0), Vector(1, 0, 0), Vector(0, 1, 0))
+        >>> frame2 = Frame(Point(1, 1, 1), Vector(0, 0, 1), Vector(0, 1, 0))
+        >>> steps = 5
+        >>> frames = frame1.interpolate_frames(frame2, steps)
+        >>> assert len(frames) == steps
+        True
+        """
+        return [self.interpolate_frame(other, t) for t in linspace(0, 1, steps)]
 
     def euler_angles(self, static=True, axes="xyz"):
         """The Euler angles from the rotation given by the frame.